Added the outside panel, it simply directs the elevator to a floor, but provides the preferred direction. Once the floor is visited it turns off the light for that direction.
My remaining concern is some edge cases when fitting this all together, for
example when getting to a floor, and changing direction, does the correct light
get turned off. However this will be easier to determine later when fitting all
the pieces together, so or now I have focused on only the obvious test cases.